feat: split app initialization into modules
