+++ js/ageSetup.js
@@ -1,0 +1,5 @@
+import { updateAge } from './age.js';
+
+export function setupAgeListener(inputs) {
+  inputs.a_dob.addEventListener('input', updateAge);
+}
